+++ src/encodables/option_result.rs
